last update
getting all yogioh cards data by scraping  the yogioh web site.
the execution of this code my take a long time in my case it took 40 minutes it my take longer or less depending on the internet speed and the cpu.
the extracted data is also present in the project folder.